requirements.txt
aiofiles==0.8.0
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.5.1
bleach==5.0.1
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
ci-info==0.2.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
configparser==5.2.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
etelemetry==0.3.0
Events==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
fitz-douyu==0.0.1
fitz-message==0.2
Flask==2.2.3
frontend==0.0.3
future==0.18.2
gitdb==4.0.9
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
looseversion==1.0.1
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nibabel==4.0.1
nipype==1.8.2
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdfreader==0.1.11
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
prov==2.0.0
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pycryptodome==3.15.0
pydeck==0.7.1
pydot==1.4.2
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.5.0
PyPDF4==1.27.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytils==0.4.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
pyxnat==1.4
pyzmq==23.2.0
rdflib==6.1.1
requests==2.28.1
rich==12.5.1
scikit-learn==1.2.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.11.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tools==0.1.9
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
traits==6.3.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.18.2
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.6.1
wincertstore==0.2
zipp==3.8.1
